feat: credentials

+++ anemone-kernel/src/fs/api/unlinkat.rs
@@ -4,18 +4,28 @@
 //! - https://www.man7.org/linux/man-pages/man2/unlinkat.2.html
 
 use anemone_abi::fs::linux::at::AT_REMOVEDIR;
+use typed_path::UnixComponent;
 
 use crate::{
     fs::api::args::AtFd,
     prelude::{
         handler::{TryFromSyscallArg, syscall_arg_flag32},
-        user_access::c_readonly_string,
+        user_access::c_readonly_path,
         *,
     },
 };
 
 struct UnlinkAtFlags {
     remove_dir: bool,
+}
+
+fn rmdir_special_last_component_error(path: &Path) -> Option<SysError> {
+    match path.components().last()? {
+        UnixComponent::RootDir => Some(SysError::Busy),
+        UnixComponent::CurDir => Some(SysError::InvalidArgument),
+        UnixComponent::ParentDir => Some(SysError::DirNotEmpty),
+        UnixComponent::Normal(_) => None,
+    }
 }
 
 impl TryFromSyscallArg for UnlinkAtFlags {
@@ -34,11 +44,17 @@ impl TryFromSyscallArg for UnlinkAtFlags {
 #[syscall(SYS_UNLINKAT)]
 fn unlinkat(
     dirfd: AtFd,
-    #[validate_with(c_readonly_string::<MAX_PATH_LEN_BYTES>)] pathname: Box<str>,
+    #[validate_with(c_readonly_path)] pathname: Box<str>,
     flags: UnlinkAtFlags,
 ) -> Result<u64, SysError> {
     let path = Path::new(pathname.as_ref());
     let task = get_current_task();
+
+    if flags.remove_dir {
+        if let Some(err) = rmdir_special_last_component_error(&path) {
+            return Err(err);
+        }
+    }
 
     let (parent, name) = if path.is_absolute() {
         task.lookup_parent_path(&path, ResolveFlags::empty())?
@@ -48,8 +64,25 @@ fn unlinkat(
     };
 
     let leaf = Path::new(name.as_str());
+    let victim = task.lookup_path_from(&parent, leaf, ResolveFlags::UNFOLLOW_LAST_SYMLINK)?;
+
+    parent.mount().ensure_writable()?;
+    let checker = FsPermChecker::for_current_fs();
+    checker.check_path(&parent, FsAccess::WRITE | FsAccess::EXECUTE)?;
+    if parent.inode().perm().contains(InodePerm::ISVTX)
+        && !checker.is_owner(victim.inode())
+        && !checker.is_owner(parent.inode())
+        && !checker.has_cap(Capability::FOWNER)
+    {
+        return Err(SysError::PermissionDenied);
+    }
+
+    let leaf = Path::new(name.as_str());
     if flags.remove_dir {
-        vfs_rmdir_at(&parent, leaf)?;
+        vfs_rmdir_at(
+            &parent,
+            PathResolution::new(leaf, ResolveFlags::UNFOLLOW_LAST_SYMLINK),
+        )?;
     } else {
         vfs_unlink_at(&parent, leaf)?;
     }
